fix: refactor base routes for editor refactor routes
