+++ frontend/src/app/last-will/editor/final/page.tsx
@@ -1,7 +1,7 @@
 'use client'
-import { Route } from '../../../../../components/ButtonsAndLinks/Route/Route'
-import { Headline } from '../../../../../components/Headline/Headline'
-import { routes } from '../../../../../services/routes/routes'
+import { Route } from '../../../../components/ButtonsAndLinks/Route/Route'
+import { Headline } from '../../../../components/Headline/Headline'
+import { routes } from '../../../../services/routes/routes'
 
 /**
  * Final Page for copy last will.
