+++ frontend/src/app/last-will/editor/layout.tsx
