+++ frontend/src/app/last-will/editor/succession/page.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { Headline } from '../../../../../components/Headline/Headline'
+import { Headline } from '../../../../components/Headline/Headline'
 
 /**
  * Succession Page
