+++ frontend/src/app/last-will/editor/testator/page.tsx
@@ -1,9 +1,9 @@
 'use client'
-import { Button } from '../../../../../components/ButtonsAndLinks/Button/Button'
-import { Route } from '../../../../../components/ButtonsAndLinks/Route/Route'
-import { Headline } from '../../../../../components/Headline/Headline'
-import { routes } from '../../../../../services/routes/routes'
-import { useLastWillContext } from '../../../../../store/last-will/LastWillContext'
+import { Button } from '../../../../components/ButtonsAndLinks/Button/Button'
+import { Route } from '../../../../components/ButtonsAndLinks/Route/Route'
+import { Headline } from '../../../../components/Headline/Headline'
+import { routes } from '../../../../services/routes/routes'
+import { useLastWillContext } from '../../../../store/last-will/LastWillContext'
 
 /**
  * Testator Page
